+++ app/src/main/java/com/example/angelica/cameratest/GalleryImageItemRecyclerViewAdapter.java
@@ -6,7 +6,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-
+import android.widget.Toast;
 import com.example.angelica.cameratest.GalleryImageItemFragment.OnListFragmentInteractionListener;
 
 import java.util.ArrayList;
@@ -14,13 +14,13 @@ import java.util.ArrayList;
 
 public class GalleryImageItemRecyclerViewAdapter extends RecyclerView.Adapter<GalleryImageItemRecyclerViewAdapter.ViewHolder> {
 
-    private final OnListFragmentInteractionListener mListener;
+    private final GalleryImageItemFragment.OnListFragmentInteractionListener mListener;
 
     private ArrayList<LoadedImage> photos = new ArrayList<LoadedImage>();
     Context context;
 
 
-    public GalleryImageItemRecyclerViewAdapter(ArrayList<LoadedImage> photos, Context context, OnListFragmentInteractionListener listener) {
+    public GalleryImageItemRecyclerViewAdapter(ArrayList<LoadedImage> photos, Context context, GalleryImageItemFragment.OnListFragmentInteractionListener listener) {
         super();
         this.photos = photos;
         this.mListener = listener;
@@ -38,17 +38,18 @@ public class GalleryImageItemRecyclerViewAdapter extends RecyclerView.Adapter<Ga
     public void onBindViewHolder(final ViewHolder holder, int position) {
         holder.getImageData = photos.get(position);
         holder.mIdView.setImageBitmap(photos.get(position).getBitmap());
-
-        holder.mView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (null != mListener) {
-                    // Notify the active callbacks interface (the activity, if the
-                    // fragment is attached to one) that an item has been selected.
-                    mListener.onListFragmentInteraction(holder.getImageData);
-                }
-            }
-        });
+        final String message = String.valueOf(position);
+//        holder.mIdView.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View v) {
+//                if (null != mListener) {
+//                    // Notify the active callbacks interface (the activity, if the
+//                    // fragment is attached to one) that an item has been selected.
+//                    mListener.onListFragmentInteraction(holder.getImageData);
+//                    Toast.makeText(v.getContext(),holder.getImageData.getPath(), Toast.LENGTH_SHORT).show();
+//                }
+//            }
+//        });
     }
 
     @Override
@@ -56,7 +57,7 @@ public class GalleryImageItemRecyclerViewAdapter extends RecyclerView.Adapter<Ga
         return photos.size();
     }
 
-    public class ViewHolder extends RecyclerView.ViewHolder {
+    public class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
         public final View mView;
         public final ImageView mIdView;
         //public final TextView mContentView;
@@ -66,7 +67,23 @@ public class GalleryImageItemRecyclerViewAdapter extends RecyclerView.Adapter<Ga
             super(view);
             mView = view;
             mIdView = (ImageView) view.findViewById(R.id.id_camera_image);
+            view.setOnClickListener(this);
             //mContentView = (TextView) view.findViewById(R.id.content);
+        }
+
+
+        @Override
+        public void onClick(View view) {
+            final int position = getAdapterPosition();
+            if(position != RecyclerView.NO_POSITION) {
+                System.out.println("Click " + position);
+                Toast.makeText(view.getContext(), "Click " + position + " photo " + photos.get(position).getPath(), Toast.LENGTH_SHORT).show();
+                if (null != mListener) {
+//                    // Notify the active callbacks interface (the activity, if the
+//                    // fragment is attached to one) that an item has been selected.
+                    mListener.onListFragmentInteraction(photos.get(position));
+                }
+            }
         }
 
     }
