+++ app/src/main/java/com/example/angelica/cameratest/LoadedImage.java
@@ -11,12 +11,29 @@ import android.graphics.Bitmap;
  */
 public class LoadedImage {
     Bitmap mBitmap;
+    String path;
 
-    LoadedImage(Bitmap bitmap) {
+
+    LoadedImage(Bitmap bitmap,String path) {
         mBitmap = bitmap;
+        this.path = path;
+
     }
 
     public Bitmap getBitmap() {
         return mBitmap;
     }
+
+    public void setBitmap(Bitmap mBitmap) {
+        this.mBitmap = mBitmap;
+    }
+
+    public String getPath() {
+        return path;
+    }
+
+    public void setPath(String path) {
+        this.path = path;
+    }
+
 }
